requirements.txt
Django==4.2.16
django-active-link==0.1.8
django-allauth==65.3.0
django-ckeditor==6.7.0
django-cleanup==6.0.0
django-tinymce==3.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
python-dotenv==1.0.0
requests==2.32.3
sorl-thumbnail==12.9.0